requirements.txt
playwright>=1.44.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
scikit-learn>=1.4.0
numpy>=1.26.0
scipy>=1.13.0
matplotlib>=3.8.0
pytest>=8.2.0
fastapi>=0.111.0
uvicorn>=0.29.0
